Use `u32`s for indices into `[MatcherLoc]`
Needed so I can add another field to `MatcherPos` without growing it.

+++ compiler/rustc_expand/src/mbe/macro_rules.rs
@@ -363,7 +363,7 @@ pub(super) trait Tracker<'matcher> {
     fn prepare(&mut self, which_matcher: WhichMatcher, matcher: &'matcher [MatcherLoc]);
 
     /// A [`MatcherLoc`] is about to be matched against `token` at `input_pos`.
-    fn trying_match(&mut self, input_pos: u32, token: &Token, loc_index: usize);
+    fn trying_match(&mut self, input_pos: u32, token: &Token, loc_index: u32);
 
     /// A [`MatcherLoc`] successfully consumed input from the parser.
     ///
@@ -371,7 +371,7 @@ pub(super) trait Tracker<'matcher> {
     /// single tokens, when they successfully match the token at `input_pos`. It is also called for
     /// [`MatcherLoc::MetaVarDecl`] when non-terminal parsing is guaranteed to occur (i.e. after
     /// [`Parser::nonterminal_may_begin_with()`] returns `true`).
-    fn matched_one(&mut self, input_pos: u32, loc_index: usize);
+    fn matched_one(&mut self, input_pos: u32, loc_index: u32);
 
     /// This is called after an arm has been parsed, either successfully or unsuccessfully. When
     /// this is called, `before_match_loc` was called at least once (with a `MatcherLoc::Eof`).
@@ -404,9 +404,9 @@ pub(super) struct NoopTracker;
 impl<'matcher> Tracker<'matcher> for NoopTracker {
     fn prepare(&mut self, _which_matcher: WhichMatcher, _matcher: &'matcher [MatcherLoc]) {}
 
-    fn trying_match(&mut self, _input_pos: u32, _token: &Token, _loc_index: usize) {}
+    fn trying_match(&mut self, _input_pos: u32, _token: &Token, _loc_index: u32) {}
 
-    fn matched_one(&mut self, _input_pos: u32, _loc_index: usize) {}
+    fn matched_one(&mut self, _input_pos: u32, _loc_index: u32) {}
 
     fn ambiguity(&mut self) {}
